% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEdge.R
\name{addEdge}
\alias{addEdge}
\alias{addNode}
\alias{mediate}
\alias{removeEdge}
\alias{removeNode}
\alias{setPolarity}
\alias{setWeight}
\title{Functions to manipulate the nodes and edges within a causal loop diagram}
\usage{
addEdge(CLD)

addNode(CLD)

mediate(CLD)

removeEdge(CLD)

removeNode(CLD)

setPolarity(CLD, edgeIndex, from, to, polarity)

setWeight(CLD, edgeIndex, from, to, weight)
}
\arguments{
\item{CLD}{An object of class CLD (see \code{help(CLD)}}
}
\value{
xxx
}
\details{
The functions \code{setPolarity()} and \code{setWeight()}, accept either a
numeric vector of \code{edgeIndex} values or equal-length character vectors
\code{from} and \code{to}.
}
\examples{
addEdge(1)
}
\author{
Jarrod E. Dalton
}
