% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNodeGroup.R
\name{addNodeGroup}
\alias{addNodeGroup}
\title{Create a format class for a group of nodes in a causal loop diagram.}
\usage{
addNodeGroup(CLD, groups, ...)
}
\arguments{
\item{CLD}{A causal loop diagram object}

\item{groups}{A character vector indicating the group name(s)}

\item{...}{Other node attributes passed to \code{setNodeFormat()}.}
}
\value{
An updated CLD object; in particular, with updated \code{$formats$node} dataset.
}
\description{
This function specifies the node attributes for a group of nodes, allowing
  easy format management for node groups.
}
\examples{
L <- CLD(from=c("a","a","c","c","d"), to=c("b","c","a","d","a")) \%>\%
  addNodeData(tibble(node=c("c","d","e","f"), group=c("I"," ","II","I"))) \%>\%
  addNodeGroup(groups=c("I","II"), fillcolor="yellow", shape=c("box","ellipse"))
}
\seealso{
\code{\link{setNodeFormat}}
}
