% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLD.R
\name{CLD}
\alias{CLD}
\title{Create a causal loop diagram (CLD) object}
\usage{
CLD(from, to, polarity = 1, weight = 1)
}
\arguments{
\item{from}{A vector of type character, providing the variables at the beginning of
each (directional) causal link.}

\item{to}{A vector of type character, containing the variables at the end of each
(directional) causal link. Must be the same length as \code{from}.}

\item{polarity}{A numeric vector containing the edge polarities. Must be either
a scalar or a vector the same length as \code{from} and \code{to}; and must
contain values in \code{c(-1,1)}. Defaults to 1 for all edges.}

\item{weight}{A numeric vector containing the edge weights (e.g., importance
weights). Must be the same length as \code{from} and \code{to} and must contain
values in \code{1:5}. Defaults to 1 for all edges.}

\item{...}{Additional arguments passed to \code{formatEdges()}.}
}
\value{
A CLD object. This is a list with the following elements:
\itemize{
   \item{\code{edf}} An edge data frame with the following columns:
       \itemize{
           \item{\code{from}} --- Source variable for the edge.
           Class \code{character}.
           \item{\code{to}} --- Destination variable for the edge.
           Class \code{character}.
           \item{\code{polarity}} --- Edge polarities. Class \code{integer}.
           \item{\code{weight}} --- Edge weights (e.g., importance weight).
           Class \code{integer}.
       }
}
}
\description{
Create a causal loop diagram (CLD) object
}
\examples{
L <- CLD(from=c("a","a","c","c","d"), to=c("b","c","a","d","a"))
L$edges


}
