% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLD2GoogleSheet.R
\name{CLD2GoogleSheet}
\alias{CLD2GoogleSheet}
\title{Title}
\usage{
CLD2GoogleSheet(CLD, key, replace = FALSE)
}
\arguments{
\item{CLD}{A CLD object.}

\item{key}{Google Sheets key.}

\item{replace}{Logical. Replace table contents?}
}
\value{
The number 1.
}
\description{
Title
}
