% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edf2CLD.R
\name{edf2CLD}
\alias{edf2CLD}
\title{Convert edge data frame into a causal loop diagram (CLD) object}
\usage{
edf2CLD(edf, ndf, ...)
}
\arguments{
\item{edf}{An edge data frame containing at least character columns \code{from} and
\code{to}. This data frame may optionally contain columns \code{polarity} and/or
\code{weight}. See \code{help(CLD)} for restrictions on these variables. All other
variables are ignored.}

\item{ndf}{An optional node data frame containing at least a character column \code{node}.
When \code{ndf} is specified, \code{edf2CLD()} loads the node information into the
CLD object and checks the names of the variables contained in \code{edf$from} and
\code{edf$to} for membership in \code{ndf$node}. This data frame may optionally contain
a column named \code{group} to designate individual variables into groups. All other
variables are ignored.}
}
\value{
A newly-created CLD object.
}
\description{
Convert edge data frame into a causal loop diagram (CLD) object
}
\details{
The function searches \code{edf} for columns named \code{polarity} and
\code{weight}. If these are found, then these values are passed to \code{CLD()}
instead of whatever values may be passed as arguments to this function). See example.

For information on formatting nodes separately according to the \code{group} variable,
see \code{help(CLD)}.
}
\examples{
#note precedence in taking the polarity argument here
edf2CLD(tibble(from=c("a","a"), to=c("b","c"), polarity=-1), polarity=1)
\dontrun{
  edf2CLD(data.frame(from="a", to=c("b","c")))  #produces error
  edf2CLD(data.frame(from="a", to=c("b","c"), stringsAsFactors=FALSE))
}
}
\seealso{
\link{CLD()}.
}
\author{
Jarrod E. Dalton
}
