% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatEdges.R
\name{formatEdges}
\alias{formatEdges}
\alias{formatNodes}
\title{Functions to format nodes and edges in a causal loop diagram (CLD)}
\usage{
formatEdges(CLD, edge.style.pos = "solid", edge.style.neg = "dotted",
  edge.color.pos = "ForestGreen", edge.color.neg = "red",
  edge.arrowhead.pos = "vee", edge.arrowhead.neg = "box",
  edge.penwidthAdj.pos = 1, edge.penwidthAdj.neg = 1)

formatNodes(ndf)
}
\arguments{
\item{CLD}{An object of class CLD}

\item{edge.style.pos}{Style of edge lines for links with positive ("+") polarity}

\item{edge.style.neg}{Style of edge lines for links with negative ("-") polarity}

\item{edge.color.pos}{Edge color for "+" links}

\item{edge.color.neg}{Edge color for "-" links}

\item{edge.arrowhead.pos}{Arrowhead type for "+" links}

\item{edge.arrowhead.neg}{Arrowhead type for "-" links}

\item{edge.penwidthAdj.pos}{Multiplicative adjustment factor to edge pen width
for "+" links. This is calculated as \code{CLD$edf$weight *edge.penwidthAdj.pos}.}

\item{edge.penwidthAdj.neg}{Multiplicative adjustment factor to edge pen width
for "-" links. This is calculated as \code{CLD$edf$weight *edge.penwidthAdj.neg}.}
}
\value{
An updated CLD object.
}
\description{
Functions to format nodes and edges in a causal loop diagram (CLD)
}
