% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{internalFunctions}
\alias{internalFunctions}
\alias{processEdgeInput}
\alias{allEdfNodesListedInNdf}
\title{Internal functions in the causalloop package}
\usage{
processEdgeInput(from, to)

allEdfNodesListedInNdf(CLD)
}
\arguments{
\item{from}{A vector of type character, providing the variables at the beginning of
each (directional) causal link.}

\item{to}{A vector of type character, containing the variables at the end of each
(directional) causal link. Must be the same length as \code{from}.}

\item{CLD}{a causal loop diagram (CLD) object.}
}
\value{
data frame containing the following variables:
}
\description{
\code{processEdgeInput()} checks user-inputted edge specification for
  CLD manipulation functions.

  \code{allEdfNodesListedInNdf()} checks to ensure that
  all nodes listed in the \code{from} and \code{to} columns of the \code{edges} table
  within a CLD object are contained in the \code{node} column within the \code{nodes}
  table.
}
\details{
The parameters \code{from} and \code{to} must either be of equal length; or
  be of differing lengths, under the restriction that at least one of the two is of
  length 1.
}
\examples{
processEdgeInput(from=c("a","a"), to=c("b")) #duplicates are removed w/ warning
processEdgeInput(from="a", to=c("b","c","d")) #replicates as in data.frame()
processEdgeInput(from=c("b","c","d"), to="z") #replicates as in data.frame()
processEdgeInput(from=c("a","b"), to=c("b","a","c")) #produces error as in data.frame()
}
