% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CLD.R
\name{plot.CLD}
\alias{plot.CLD}
\alias{plotCLD}
\title{Plot a causal loop diagram object}
\source{
\url{http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html}\cr
  See especially the section on Attributes

  \url{http://graphviz.org/}\cr
  \url{http://graphviz.org/content/attrs}
}
\usage{
\method{plot}{CLD}(CLD, nodes = NULL, steps = 1, recolor = TRUE,
  textWidth = 10)
}
\arguments{
\item{CLD}{an object of class \code{CLD}}

\item{nodes}{Subset of nodes in the CLD to visualize.}

\item{steps}{Nodes not in the \code{nodes} parameter will also be
displayed if they are within \code{steps} edges of the CLD. By
default, connections involving these nodes are recolored
(see parameter \code{recolor}).}

\item{recolor}{Logical. Recolor ancillary links (i.e., links involving
nodes other than those specified in the \code{nodes} argument)
to \code{gray70}.}

\item{textWidth}{The target width of node names, which are all passed to
\code{stringr::str_wrap()}.}
}
\description{
Causal loop diagrams are generated through an interface to the
  \code{DiagrammeR} package. Node and edge characteristics can be modified directly
  by manipulating the CLD object or by using functions \code{setEdgeFormat()} and
  \code{setNodeFormat()}.
}
\details{
See 'Sources' for links to additional documentation from the
  \code{DiagrammeR} package and the GraphViz website.
}
\examples{
\dontrun{
#* Plots may open in a browser.
L <- CLD(from     = c("a","a","b","c","d","d","e"),
         to       = c("b","c","a","d","b","a","d"),
         polarity = c(1,1,-1,-1,1,-1,1)) \%>\%
  addNodeData(tibble::tibble(node="c", group="core")) \%>\%
  addNodeGroup("core", fontcolor="red", color="yellow")
plot(L)
plot(L, nodes=c("c"))
}
}
\author{
Jarrod Dalton
}
