% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processEdgeInput.R
\name{processEdgeInput}
\alias{processEdgeInput}
\title{Check user-inputted edge specification for CLD manipulation functions (internal)}
\usage{
processEdgeInput(from, to)
}
\arguments{
\item{from}{A vector of type character, providing the variables at the beginning of
each (directional) causal link.}

\item{to}{A vector of type character, containing the variables at the end of each
(directional) causal link. Must be the same length as \code{from}.}
}
\value{
data frame containing the following variables:
}
\description{
Check user-inputted edge specification for CLD manipulation functions (internal)
}
\details{
The parameters \code{from} and \code{to} must either be of equal length; or
be of differing lengths, under the restriction that at least one of the two is of
length 1.
}
\examples{
processEdgeInput(from=c("a","a"), to=c("b")) #duplicates are removed w/ warning
processEdgeInput(from="a", to=c("b","c","d")) #replicates as in data.frame()
processEdgeInput(from=c("b","c","d"), to="z") #replicates as in data.frame()
processEdgeInput(from=c("a","b"), to=c("b","a","c")) #produces error as in data.frame()
}
