% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEdgeFormat.R
\name{setEdgeFormat}
\alias{setEdgeFormat}
\alias{setNodeFormat}
\title{Functions to format nodes and edges in a causal loop diagram (CLD)}
\usage{
setEdgeFormat(CLD, property, value, polarity)

setNodeFormat(CLD, groups, ...)
}
\arguments{
\item{CLD}{An object of class CLD}

\item{property}{The property of the edge or node being manipulated. Allowable values
are provided in the details.}

\item{value}{Value of the parameter to be assigned. Note that this is not validated
against what values are allowed by GraphViz. See the documentation for GraphViz
for a list of allowable values.}

\item{polarity}{Polarity of edges for which the edge format is being manipulated.
When this parameter is not specified, \code{setEdgeFormat()} will apply the format
change to both positive and negative links (with warning).}

\item{...}{Format parameters as described in the details section below.}

\item{group}{Character vector of group label(s) of nodes for which the node
format is being manipulated.  When this parameter is not specified,
\code{setNodeFormat()} applies the formatting to all node groups that are
currently included in the \code{CLD} object (under \code{$formats$node$group}).}
}
\value{
An updated CLD object.
}
\description{
Functions to format nodes and edges in a causal loop diagram (CLD)
}
\details{
For \code{setEdgeFormat()}, the following parameters are acted upon (all
others are ignored):
\itemize{
   \item{\code{style}} --- Line style (e.g., \code{"solid"} or \code{"dotted"})
   \item{\code{color}} --- Line color
   \item{\code{arrowhead}} --- Arrowhead style (e.g., \code{"vee"}, \code{"normal"},
   \code{"box"}, or \code{"diamond"}).
   \item{\code{penwidthAdj}} ---  Pen width adjustment factor. This gets multiplied
   by the edge weight to establish the thickness ("pen width") of each line.
}
For \code{setNodeFormat()}, the following parameters are acted upon (all
others are ignored):
\itemize{
   \item{\code{shape}} --- Node shape (e.g., \code{"rectangle"}, \code{"oval"},
   \code{"box"} or \code{"egg"})
   \item{\code{style}} --- Node line style (e.g., \code{"solid"}, \code{"dashed"},
   \code{"dotted"} or \code{"bold"})
   \item{\code{color}} --- Line color for the node shape
   \item{\code{fillcolor}} --- Fill color for the node
   \item{\code{fontname}} --- Font for the node label
   \item{\code{fontcolor}} --- Font color for the node label
   \item{\code{height}} --- Base height of the node (all graphs are plotted with
   \code{fixedsize = FALSE}, such that height and width change to accomodate the
   node labels)
   \item{\code{width}} --- Base width of the node (all graphs are plotted with
   \code{fixedsize = FALSE}, such that height and width change to accomodate the
   node labels)
}
Default attributes are adopted unless the attributes are specified through the
function.
}
\examples{
L <- CLD(from=c("a","a","c"), to=c("b","c","a")) \%>\%
  setEdgeFormat("color", "midnightblue", polarity=-1) \%>\%
  setNodeFormat("fillcolor", "yellow")
#plot(L)
}
