% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEdge.R
\name{addEdge}
\alias{addEdge}
\alias{addNode}
\alias{mediate}
\alias{removeEdge}
\alias{removeNode}
\alias{setPolarity}
\alias{setWeight}
\title{Functions to manipulate the nodes and edges within a causal loop diagram}
\usage{
addEdge(CLD)

addNode(CLD)

mediate(CLD)

removeEdge(CLD)

removeNode(CLD)

setPolarity(CLD, edgeIndex, from, to, polarity)

setWeight(CLD, edgeIndex, from, to, weight, ...)
}
\arguments{
\item{CLD}{An object of class CLD (see \code{help(CLD)}}

\item{edgeIndex}{A numeric vector of edge indices from the CLD object.}

\item{from}{A vector of type character, providing the variables at the beginning of
each (directional) causal link.}

\item{to}{A vector of type character, containing the variables at the end of each
(directional) causal link. Must be the same length as \code{from}.}

\item{polarity}{A numeric vector containing the edge polarities. Must be the same
length as \code{from} and \code{to} and must contain values in \code{c(-1,0,1)}.
Defaults to 1 for all edges.}

\item{weight}{A numeric vector containing the edge weights (e.g., importance
weights). Must be the same length as \code{from} and \code{to} and must contain
values in \code{1:5}. Defaults to 1 for all edges.}

\item{...}{additional arguments to \code{formatEdges()} or \code{formatNodes}. See details.}
}
\value{
An updated CLD object.
}
\details{
The functions \code{setPolarity()} and \code{setWeight()}, accept either a
numeric vector of \code{edgeIndex} values or equal-length character vectors
\code{from} and \code{to}. Polarity values of 0 represent situations where the
direction of the relationship may be different for different subgroups.
}
\examples{
addEdge(1)
}
\author{
Jarrod E. Dalton
}
