% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoops.R
\name{getLoops}
\alias{getLoops}
\title{Extract Feedback Loops from a Causal Loop Diagram}
\usage{
getLoops(CLD)
}
\arguments{
\item{CLD}{a CLD object}
}
\value{
a list of 2-element lists, each with element \code{[[1]]} containing
  a feedback loop and element \code{[[2]]} containing the loop's type (i.e.,
  either "balancing" or "reinforcing"). Balancing loops have an odd number
  of links with negative polarity, and reinforcing loops have an even
  number of links with negative polarity.
}
\description{
Extract Feedback Loops from a Causal Loop Diagram
}
\details{
This function applies Johnson's algorithm for finding cycles
  in a directed graph. In particular, we adapt for the \code{causalloop}
  package Jacobien Carstens' implementation of Johnson's algorithm,
  which can be found on GitHub at the link below (see References)
}
\examples{
1:3
}
\references{
\url{https://github.com/queenBNE/DirectedGraphsCycles}
}
