% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CLD.R
\name{plot.CLD}
\alias{plot.CLD}
\title{Plot a causal loop diagram object}
\source{
\url{http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html}\cr
  See especially the section on Attributes

  \url{http://graphviz.org/}\cr
  \url{http://graphviz.org/content/attrs}
}
\usage{
\method{plot}{HydeNetwork}(x, nodes, steps = 1, ...)
}
\arguments{
\item{x}{an object of class \code{HydeNetwork}}

\item{nodes}{(optional) character vector containing the nodes to be plotted. By default,
when this parameter is specified, all nodes having links to the specified nodes are
also visualized (i.e., \code{steps = 1}).
between nodes.}

\item{steps}{(optional) Maximum number of steps away from the specified \code{nodes} to
be included in the figure. Defaults to 1. Only has an effect when \code{nodes} is
specified.}

\item{...}{for the \code{plot} method, additional arguments to be passed to
\code{DiagrammeR::render_graph}.}
}
\description{
Causal loop diagrams are generated through an interface to the
  \code{DiagrammeR} package. Node and edge characteristics can be modified directly
  by manipulating the CLD object or by using functions \code{setEdgeFormat()} and
  \code{setNodeFormat()}.
}
\details{
See 'Sources' for links to additional documentation from the \code{DiagrammeR}
  package and the GraphViz website.
}
\author{
Jarrod Dalton
}
